# fix (sdk): interpolate, replace same var prefix