--- a/sdk/interpolate/interpolate_test.go
+++ b/sdk/interpolate/interpolate_test.go
@@ -224,8 +224,7 @@ func TestDo(t *testing.T) {
 			name: "same prefix",
 			args: args{
 				input: `{"HOST": "customer{{.cds.env.lb.prefix}}.{{.cds.env.lb}}"}`,
-				//vars:  map[string]string{"cds.env.lb": "lb", "cds.env.lb.prefix": "myprefix"},
-				vars: map[string]string{"cds.env.lb": "lb", "cds.env.lb.prefix": "myprefix"},
+				vars:  map[string]string{"cds.env.lb": "lb", "cds.env.lb.prefix": "myprefix"},
 			},
 			want: `{"HOST": "customermyprefix.lb"}`,
 		},