close #2046

before fix:
Running tool: /Users/yesnault/go/bin/go test -benchmem -run=^$ github.com/ovh/cds/sdk/interpolate -bench ^BenchmarkDo$

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
BenchmarkDo-4   	   10000	    170029 ns/op	   53574 B/op	     546 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	1.732s
Success: Benchmarks passed.

Running tool: /Users/yesnault/go/bin/go test -benchmem -run=^$ github.com/ovh/cds/sdk/interpolate -bench ^BenchmarkDoNothing$

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
BenchmarkDoNothing-4   	   50000	     30224 ns/op	   11906 B/op	     247 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	1.798s
Success: Benchmarks passed.

---
With fix:

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
BenchmarkDo-4   	   10000	    183572 ns/op	   55689 B/op	     556 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	1.867s
Success: Benchmarks passed.

Running tool: /Users/yesnault/go/bin/go test -benchmem -run=^$ github.com/ovh/cds/sdk/interpolate -bench ^BenchmarkDoNothing$

goos: darwin
goarch: amd64
pkg: github.com/ovh/cds/sdk/interpolate
BenchmarkDoNothing-4   	   50000	     30278 ns/op	   11904 B/op	     247 allocs/op
PASS
ok  	github.com/ovh/cds/sdk/interpolate	1.825s
Success: Benchmarks passed.

-----

Signed-off-by: Yvonnick Esnault <yvonnick.esnault@corp.ovh.com>